Added verifications

+++ src/controllers/UserController.js
@@ -10,6 +10,12 @@ module.exports = {
     async create(request, response) {
         const { name, email, phone, location } = request.body;
 
+        const user = await connection('users').where({ email }).first();
+
+        if (user) {
+            return response.status(401).send({ error: 'E-mail already exists.' });
+        }
+
         await connection('users').insert({
             name,
             email,
@@ -25,6 +31,12 @@ module.exports = {
 
         const { name, email, phone, location } = request.body;
 
+        const updateUser = await connection('users').where({ id }).first();
+
+        if (!updateUser) {
+            return response.status(404).send({ error: 'User not found.' });
+        }
+
         await connection('users')
             .where({ id })
             .update({ name, email, phone, location });
@@ -34,6 +46,12 @@ module.exports = {
 
     async delete(request, response) {
         const { id } = request.params;
+
+        const deleteUser = await connection('users').where({ id }).first();
+
+        if (!deleteUser) {
+            return response.status(404).send({ error: 'User not found.' });
+        }
 
         await connection('users').where({ id }).delete();
